Development database, pagination and item listing
The backend has been configured to stop using the PopulateDatabase utility and is connecting to a development database instead, using a hibernate.ddl-auto validate property instead of update.

Support for paginated repositories has been added, as well as a POJO class that is passed to AngularJS in order to support pagination in frontend.

A draft item listing functionality has been added in order to test the former.